@@ -1,28 +1,34 @@
-package com.francis.week6
+package com.francis.week6.ui
 
 import android.Manifest
 import android.content.Intent
 import android.content.pm.PackageManager
 import android.net.Uri
-import android.os.Build
 import android.os.Bundle
-import android.view.*
+import android.view.LayoutInflater
+import android.view.MenuItem
+import android.view.View
+import android.view.ViewGroup
 import android.widget.FrameLayout
 import android.widget.TextView
+import androidx.activity.result.contract.ActivityResultContracts
 import androidx.appcompat.app.AlertDialog
 import androidx.appcompat.app.AppCompatActivity
 import androidx.appcompat.widget.Toolbar
 import androidx.cardview.widget.CardView
-import androidx.core.content.ContextCompat
 import androidx.fragment.app.Fragment
 import androidx.navigation.findNavController
+import androidx.navigation.fragment.findNavController
 import androidx.recyclerview.widget.DividerItemDecoration
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
+import com.francis.week6.R
+import com.francis.week6.adapter.PhoneDetailsAdapter
 import com.francis.week6.models.Contact
 import com.francis.week6.models.ContactStore
 import com.google.android.material.appbar.CollapsingToolbarLayout
 import com.google.android.material.floatingactionbutton.FloatingActionButton
+import java.util.*
 
 
 /**
@@ -30,6 +36,7 @@ import com.google.android.material.floatingactionbutton.FloatingActionButton
  */
 class ContactDetailsFragment : Fragment() {
 
+    private var screen: Int = 0
     private lateinit var contact: Contact
     private lateinit var fullName: String
     private lateinit var initials: String
@@ -37,7 +44,14 @@ class ContactDetailsFragment : Fragment() {
     private lateinit var initialsTextView: TextView
     private lateinit var recyclerView: RecyclerView
 
-    override fun onOptionsItemSelected(item: MenuItem): Boolean {
+    private val permissionLauncher =
+        registerForActivityResult(ActivityResultContracts.RequestPermission()) { granted ->
+            if (granted) {
+                makeCall()
+            }
+        }
+
+    private fun setMenuItemClickListener(item: MenuItem): Boolean {
         when (item.itemId) {
             android.R.id.home -> {
                 view?.findNavController()?.popBackStack() //Navigate back from fragment.
@@ -54,7 +68,7 @@ class ContactDetailsFragment : Fragment() {
                 val alert = AlertDialog.Builder(requireContext())
                 alert.setTitle("Delete")
                 alert.setMessage("Are you sure you want to delete contact?")
-                alert.setPositiveButton("Yes"){ it, _ ->
+                alert.setPositiveButton("Yes") { it, _ ->
                     ContactStore.deleteContact(contact.id!!)
                     it.dismiss()
                     view?.findNavController()?.popBackStack()
@@ -66,24 +80,16 @@ class ContactDetailsFragment : Fragment() {
         return super.onOptionsItemSelected(item)
     }
 
-    override fun onCreateOptionsMenu(menu: Menu, inflater: MenuInflater) {
-        inflater.inflate(R.menu.contact_menu, menu)
-        //Details fragment should not display save menu item.
-        menu.findItem(R.id.action_save).isVisible = false
-        super.onCreateOptionsMenu(menu, inflater)
-    }
-
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
 
-        /*
-            * Get arguments passed from navigation component
-            */
+        //Get arguments passed from navigation component
         arguments?.let {
             val args = ContactDetailsFragmentArgs.fromBundle(it)
-            contact = args.contactData
+            screen = args.screen
+            contact = args.contact
             fullName = contact.fullName?.trim().toString()
-            initials = fullName.first().toUpperCase().toString()
+            initials = fullName.first().uppercaseChar().toString()
         }
     }
 
@@ -92,76 +98,65 @@ class ContactDetailsFragment : Fragment() {
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        setHasOptionsMenu(true)
-        return inflater.inflate(R.layout.fragment_contact_details, container, false)
+        val view = inflater.inflate(R.layout.fragment_contact_details, container, false)
+        setupToolsBar(view)
+        return view
+    }
+
+    private fun setupToolsBar(view: View) {
+        (requireActivity() as AppCompatActivity).supportActionBar?.hide()
+        val toolbar = view.findViewById<Toolbar>(R.id.toolsBar)
+        toolbar.inflateMenu(R.menu.contact_menu)
+        toolbar.menu.findItem(R.id.action_save).isVisible = false
+        toolbar.setNavigationOnClickListener {
+            findNavController().popBackStack()
+        }
+        toolbar.setOnMenuItemClickListener {
+            setMenuItemClickListener(it)
+        }
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-
-        val toolbar: Toolbar = view.findViewById(R.id.toolsBar)
-        val context = (activity as AppCompatActivity?)!!
-        context.setSupportActionBar(toolbar)
 
         //Initialize edit button and setOnclick listener to it
         view.findViewById<FloatingActionButton>(R.id.edit_button).apply {
             setOnClickListener {
                 //Get fragment navigation action and navigate to next fragment.
                 val action = ContactDetailsFragmentDirections
-                    .actionContactDetailsFragmentToAddContactFragment("Edit", contact)
+                    .actionContactDetailsFragmentToAddContactFragment(
+                        if (screen == 0) "Edit" else "Add",
+                        contact
+                    )
                 view.findNavController().navigate(action)
             }
         }
 
-        /*
-         * Initialize toolbar layout and set title to full name.
-         */
+        //Initialize toolbar layout and set title to full name.
         view.findViewById<CollapsingToolbarLayout>(R.id.tools_layout).apply {
             toolBar = this
             title = fullName
         }
 
-        /*
-         * Initialize circular card view and set background color
-         */
+        //Initialize circular card view and set background color
         view.findViewById<CardView>(R.id.card_view).apply {
             setCardBackgroundColor(contact.color)
         }
 
-        /*
-         * Initialize Initials text view and set text to the contact initials
-         */
+        //Initialize Initials text view and set text to the contact initials
         view.findViewById<TextView>(R.id.detail_screen_initials_text_view).apply {
             initialsTextView = this
             text = initials
         }
 
-        /*
-         * Initialize call button and set onClick Listener
-         */
+        //Initialize call button and set onClick Listener
         view.findViewById<FrameLayout>(R.id.call_action).apply {
             setOnClickListener {
-                if (Build.VERSION.SDK_INT > 22) {
-                    //Check if permission is granted to make call else request permission.
-                    if (ContextCompat.checkSelfPermission(
-                            requireContext(),
-                            Manifest.permission.CALL_PHONE
-                        ) != PackageManager.PERMISSION_GRANTED
-                    ) {
-                        requestPermissions(
-                            arrayOf(Manifest.permission.CALL_PHONE),
-                            101
-                        )
-                    }else {
-                        makeCall()
-                    }
-                }
+                permissionLauncher.launch(Manifest.permission.CALL_PHONE)
             }
         }
 
-        /*
-         * Initialize message button and setOnclick listener.
-         */
+        //Initialize message button and setOnclick listener.
         view.findViewById<FrameLayout>(R.id.message_action).apply {
             setOnClickListener {
                 //Create intent to send message to a provider phone number.
@@ -171,9 +166,7 @@ class ContactDetailsFragment : Fragment() {
             }
         }
 
-        /*
-         * Initialize recyclerView.
-         */
+        //Initialize recyclerView.
         view.findViewById<RecyclerView>(R.id.recycler_view_list_phone).apply {
             recyclerView = this
             val data = arrayListOf<String>()
@@ -183,21 +176,20 @@ class ContactDetailsFragment : Fragment() {
             }
 
             //Set layout manager for the recyclerview.
-            layoutManager = LinearLayoutManager(requireContext(), LinearLayoutManager.VERTICAL, false)
+            layoutManager =
+                LinearLayoutManager(requireContext(), LinearLayoutManager.VERTICAL, false)
             adapter = PhoneDetailsAdapter(data) // Set recycler view adapter
 
             //add item separator to recycler view items
             addItemDecoration(
                 DividerItemDecoration(
-                    getContext(),
+                    context,
                     DividerItemDecoration.VERTICAL
                 )
             )
         }
 
-        /*
-         * Initialize email button and setOnclick listener.
-         */
+        //Initialize email button and setOnclick listener.
         view.findViewById<FrameLayout>(R.id.email_action).apply {
             //Check if contact has an email else remove email button from screen.
             if (contact.email == null || contact.email!!.isBlank()) {
@@ -214,11 +206,12 @@ class ContactDetailsFragment : Fragment() {
             }
         }
 
-        ContactStore.contact.observe(viewLifecycleOwner, {contact ->
+        ContactStore.contact.observe(viewLifecycleOwner) { contact ->
 
             if (this.contact == contact) {
                 toolBar.title = contact.fullName
-                initialsTextView.text = contact.fullName?.trim()?.toUpperCase()?.first().toString()
+                initialsTextView.text =
+                    contact.fullName?.trim()?.uppercase(Locale.getDefault())?.first().toString()
                 val data = arrayListOf<String>()
                 contact.phone?.let { data.add(it) }
                 if (contact.email?.isNotEmpty() == true) {
@@ -226,7 +219,7 @@ class ContactDetailsFragment : Fragment() {
                 }
                 recyclerView.adapter = PhoneDetailsAdapter(data)
             }
-        })
+        }
     }
 
     override fun onRequestPermissionsResult(
