@@ -1,12 +1,14 @@
-package com.francis.week6
+package com.francis.week6.adapter
 
 import android.view.View
 import android.view.ViewGroup
 import android.widget.TextView
 import androidx.cardview.widget.CardView
 import androidx.recyclerview.widget.RecyclerView
+import com.francis.week6.R
 import com.francis.week6.models.Contact
 import com.francis.week6.models.inflate
+import com.francis.week6.utils.Validator
 
 class ContactAdapter(
     private val clickListener: OnItemClickListener
