@@ -1,17 +1,21 @@
-package com.francis.week6
+package com.francis.week6.adapter
 
 import android.view.View
 import android.view.ViewGroup
 import android.widget.TextView
 import androidx.cardview.widget.CardView
+import androidx.recyclerview.widget.DiffUtil
+import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
+import com.francis.week6.R
 import com.francis.week6.models.Contact
 import com.francis.week6.models.inflate
 
-class Imp2Adapter(private val contacts: ArrayList<Contact>): RecyclerView.Adapter<Imp2Adapter.ViewHolder>() {
+class PhoneContactsAdapter(private val callback: (Contact) -> Unit) :
+    ListAdapter<Contact, PhoneContactsAdapter.ViewHolder>(DiffCallback()) {
 
-    inner class ViewHolder(itemView: View): RecyclerView.ViewHolder(itemView){
-        fun bind(contact: Contact){
+    inner class ViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView) {
+        fun bind(contact: Contact) {
             //Initialize circular card view and set the color
             itemView.findViewById<CardView>(R.id.cardView).apply {
                 setCardBackgroundColor(contact.color)
@@ -31,6 +35,8 @@ class Imp2Adapter(private val contacts: ArrayList<Contact>): RecyclerView.Adapte
             itemView.findViewById<TextView>(R.id.imp2_phone_text_view).also {
                 it.text = contact.phone
             }
+
+            itemView.setOnClickListener { callback(contact) }
         }
     }
 
@@ -39,8 +45,18 @@ class Imp2Adapter(private val contacts: ArrayList<Contact>): RecyclerView.Adapte
     }
 
     override fun onBindViewHolder(holder: ViewHolder, position: Int) {
-        holder.bind(contacts[position])
+        holder.bind(getItem(position))
     }
 
-    override fun getItemCount(): Int = contacts.size
+    override fun getItemCount(): Int = currentList.size
+
+    class DiffCallback : DiffUtil.ItemCallback<Contact>() {
+        override fun areItemsTheSame(oldItem: Contact, newItem: Contact): Boolean {
+            return oldItem === newItem
+        }
+
+        override fun areContentsTheSame(oldItem: Contact, newItem: Contact): Boolean {
+            return oldItem == newItem
+        }
+    }
 }
