@@ -1,15 +1,14 @@
 package com.francis.week6.models
 
 import android.graphics.Color
-import android.os.Build
-import android.os.Parcel
 import android.os.Parcelable
-import androidx.annotation.RequiresApi
 import com.google.firebase.database.Exclude
+import kotlinx.parcelize.Parcelize
 
 /**
  * Models Each contact object.
  */
+@Parcelize
 data class Contact(
     @get: Exclude //Exclude database from saving id field
     var id: String? = null,
@@ -21,42 +20,7 @@ data class Contact(
     var phone: String? = null,
     var email: String? = null,
     var color: Int = Color.argb(255, 156, 39, 176)
-): Parcelable{
-
-    @RequiresApi(Build.VERSION_CODES.Q)
-    constructor(parcel: Parcel) : this(
-        parcel.readString(),
-        parcel.readBoolean(),
-        parcel.readString(),
-        parcel.readString(),
-        parcel.readString(),
-        parcel.readInt()
-    )
-
-    @RequiresApi(Build.VERSION_CODES.Q)
-    override fun writeToParcel(parcel: Parcel, flags: Int) {
-        parcel.writeString(id)
-        parcel.writeBoolean(isDeleted!!)
-        parcel.writeString(fullName)
-        parcel.writeString(phone)
-        parcel.writeString(email)
-        parcel.writeInt(color)
-    }
-
-    override fun describeContents(): Int {
-        return 0
-    }
-
-    companion object CREATOR : Parcelable.Creator<Contact> {
-        @RequiresApi(Build.VERSION_CODES.Q)
-        override fun createFromParcel(parcel: Parcel): Contact {
-            return Contact(parcel)
-        }
-
-        override fun newArray(size: Int): Array<Contact?> {
-            return arrayOfNulls(size)
-        }
-    }
+) : Parcelable {
 
     override fun equals(other: Any?): Boolean {
         return if (other is Contact){
@@ -69,6 +33,4 @@ data class Contact(
         result = 31 * result + (fullName?.hashCode() ?: 0)
         return result
     }
-
-
 }
